update file types listed for SinglePlateDilutionNabDataHandler and SinglePlateNabDataHandler Consolidate duplicated schema generation method
