+++ nab/src/org/labkey/nab/multiplate/HighThroughputNabAssayProvider.java
@@ -15,21 +15,34 @@
  */
 package org.labkey.nab.multiplate;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.labkey.api.assay.plate.PlateBasedRunCreator;
 import org.labkey.api.assay.dilution.DilutionDataHandler;
+import org.labkey.api.assay.query.ResultsQueryView;
+import org.labkey.api.assay.query.RunListQueryView;
+import org.labkey.api.data.Container;
+import org.labkey.api.exp.api.ExpProtocol;
 import org.labkey.api.exp.property.Domain;
 import org.labkey.api.assay.actions.AssayRunUploadForm;
 import org.labkey.api.assay.AssayDataType;
 import org.labkey.api.assay.AssayRunCreator;
+import org.labkey.api.query.QuerySettings;
+import org.labkey.api.security.User;
 import org.labkey.api.study.assay.ParticipantVisitResolverType;
 import org.labkey.api.study.assay.SampleMetadataInputFormat;
 import org.labkey.api.study.assay.ThawListResolverType;
 import org.labkey.api.view.HtmlView;
 import org.labkey.api.view.HttpView;
+import org.labkey.api.view.ViewContext;
 import org.labkey.nab.NabAssayProvider;
+import org.labkey.nab.query.NabProtocolSchema;
+import org.springframework.validation.BindException;
 
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * User: brittp
@@ -94,9 +107,48 @@ public abstract class HighThroughputNabAssayProvider extends NabAssayProvider
     }
 
     @Override
-    public AssayRunCreator getRunCreator()
+    public AssayRunCreator<?> getRunCreator()
     {
-        return new PlateBasedRunCreator(this);
+        return new PlateBasedRunCreator<>(this);
     }
 
+
+    protected NabProtocolSchema generateAlternateProtocolSchema(User user, Container container, @NotNull ExpProtocol protocol, @Nullable Container targetStudy)
+    {
+        return new NabProtocolSchema(user, container, this, protocol, targetStudy)
+        {
+            final Map<String, Object> _extraParams = new HashMap<>();
+
+            @Override
+            protected RunListQueryView createRunsQueryView(ViewContext context, QuerySettings settings, BindException errors)
+            {
+                NabRunListQueryView queryView = new NabRunListQueryView(this, settings);
+                queryView.setExtraDetailsUrlParams(getDetailUrlParams());
+
+                return queryView;
+            }
+
+            @Override
+            protected ResultsQueryView createDataQueryView(ViewContext context, QuerySettings settings, BindException errors)
+            {
+                NabResultsQueryView queryView = new NabResultsQueryView(getProtocol(), context, settings);
+                queryView.setExtraDetailsUrlParams(getDetailUrlParams());
+
+                return queryView;
+            }
+
+            private Map<String, Object> getDetailUrlParams()
+            {
+                if (_extraParams.isEmpty())
+                {
+                    _extraParams.put("maxSamplesPerGraph", 20);
+                    _extraParams.put("graphWidth", 600);
+                    _extraParams.put("graphHeight", 550);
+                    _extraParams.put("graphsPerRow", 1);
+                    _extraParams.put("sampleNoun", "Virus");
+                }
+                return _extraParams;
+            }
+        };
+    }
 }
