+++ nab/src/org/labkey/nab/multiplate/SinglePlateDilutionNabAssayProvider.java
@@ -98,7 +98,7 @@ public class SinglePlateDilutionNabAssayProvider extends HighThroughputNabAssayP
         Map<String, Set<String>> domainMap = super.getRequiredDomainProperties();
 
         if (!domainMap.containsKey(NabAssayProvider.ASSAY_DOMAIN_SAMPLE_WELLGROUP))
-            domainMap.put(NabAssayProvider.ASSAY_DOMAIN_SAMPLE_WELLGROUP, new HashSet<String>());
+            domainMap.put(NabAssayProvider.ASSAY_DOMAIN_SAMPLE_WELLGROUP, new HashSet<>());
 
         Set<String> sampleProperties = domainMap.get(NabAssayProvider.ASSAY_DOMAIN_SAMPLE_WELLGROUP);
 
@@ -119,41 +119,7 @@ public class SinglePlateDilutionNabAssayProvider extends HighThroughputNabAssayP
     @Override
     public NabProtocolSchema createProtocolSchema(User user, Container container, @NotNull ExpProtocol protocol, @Nullable Container targetStudy)
     {
-        return new NabProtocolSchema(user, container, this, protocol, targetStudy)
-        {
-            Map<String, Object> _extraParams = new HashMap<>();
-
-            @Override
-            protected RunListQueryView createRunsQueryView(ViewContext context, QuerySettings settings, BindException errors)
-            {
-                NabRunListQueryView queryView = new NabRunListQueryView(this, settings);
-                queryView.setExtraDetailsUrlParams(getDetailUrlParams());
-
-                return queryView;
-            }
-
-            @Override
-            protected ResultsQueryView createDataQueryView(ViewContext context, QuerySettings settings, BindException errors)
-            {
-                NabResultsQueryView queryView = new NabResultsQueryView(getProtocol(), context, settings);
-                queryView.setExtraDetailsUrlParams(getDetailUrlParams());
-
-                return queryView;
-            }
-
-            private Map<String, Object> getDetailUrlParams()
-            {
-                if (_extraParams.isEmpty())
-                {
-                    _extraParams.put("maxSamplesPerGraph", 20);
-                    _extraParams.put("graphWidth", 600);
-                    _extraParams.put("graphHeight", 550);
-                    _extraParams.put("graphsPerRow", 1);
-                    _extraParams.put("sampleNoun", "Virus");
-                }
-                return _extraParams;
-            }
-        };
+        return generateAlternateProtocolSchema(user, container, protocol, targetStudy);
     }
 
     @Override
